pyproject.toml
[project]
name = "1-3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.11"
dependencies = [
    "docx2txt>=0.9",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.117.1",
    "numpy>=2.3.3",
    "openai>=1.109.0",
    "pypdf>=6.1.0",
    "pypdf2>=3.0.1",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.1.1",
    "torch>=2.8.0",
    "uvicorn>=0.37.0",
]
